requirements.txt
nltk == 3.2.5
GitPython >= 2.1.9
PyYAML >= 3.12